Replace more internal <a> tag links with TrackedLink
general refactoring
