+++ src/components/Header/HamburgerMenu.js
@@ -5,7 +5,7 @@ import { links } from './Guide'
 import { GlobalConsumer } from '../../GlobalState'
 import Button from '../Forms/Button'
 import SignInButton from './SignInButton'
-import { isExternal } from '../../utils/links'
+import TrackedLink from '../Links/TrackedLink'
 
 const HamburgerMenuContainer = styled('div')`
   display: flex;
@@ -38,9 +38,13 @@ function HamburgerMenu({ isMenuOpen }) {
           return (
             <>
               {wallet.url && (
-                <a href={wallet.url} target="_blank" rel="noopener noreferrer">
+                <TrackedLink
+                  to={wallet.url}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                >
                   Open Wallet
-                </a>
+                </TrackedLink>
               )}
               <Link to="#" onClick={signOut}>
                 Disconnect Wallet
@@ -51,13 +55,9 @@ function HamburgerMenu({ isMenuOpen }) {
       </GlobalConsumer>
       <Link to="/events">Events</Link>
       <Link to="/pricing">Pricing</Link>
-      {links.map(l =>
-        isExternal(l.url) ? (
-          <Link to={l.href}>{l.label}</Link>
-        ) : (
-          <a href={l.href}>{l.label}</a>
-        )
-      )}
+      {links.map(l => (
+        <TrackedLink to={l.href}>{l.label}</TrackedLink>
+      ))}
     </HamburgerMenuContainer>
   )
 }
